DRY'ed up all the different places that have to be touch when adding a new btc-style altcoin. This consolidates most of the necessary changes to one place when adding a new coin to ease maintainence and support.
This refactor also has the nice side-effect of allowing new coin
support to be implementable ahead-of-time into the extension
before there is firmware level support. In this case, if the
firmware doesn't support a given coin yet the client simply
won't show it.

To add a new coin search for the module enumerated with index
161 '@keepkey/device-client/dist/global/coin-type'. Within that
module find 'e.config' array which contains the list of implemented
coin profiles.

The 'FeeService' for new coins still needs to be handled seperately.
That will be refactored and consolidated in a future commit.